Sawn the classes to small parts
